move debug.cpp to address.cpp
